Moved LocalResourceRetriever to common.
