+++ dart/common/LocalResource.h
@@ -33,26 +33,42 @@
  *   ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *   POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef DART_UTILS_LOCALRESOURCERETRIEVER_H_
-#define DART_UTILS_LOCALRESOURCERETRIEVER_H_
-#include "dart/common/ResourceRetriever.h"
+#ifndef DART_COMMON_LOCALRESOURCE_H_
+#define DART_COMMON_LOCALRESOURCE_H_
+#include "dart/common/Resource.h"
 
 namespace dart {
-namespace utils {
+namespace common {
 
-/// \brief Retrieve local resources specified by file:// URIs.
-class LocalResourceRetriever : public virtual common::ResourceRetriever
+class LocalResource : public virtual Resource
 {
 public:
-  virtual ~LocalResourceRetriever() = default;
+  LocalResource(const std::string& _path);
+  virtual ~LocalResource();
 
-  bool exists(const std::string& _uri) override;
-  common::ResourcePtr retrieve(const std::string& _uri) override;
+  LocalResource(const LocalResource& _other) = delete;
+  LocalResource& operator=(const LocalResource& _other) = delete;
+
+  /// Return if the resource is open and in a valid state.
+  bool isGood() const;
+
+  // Documentation inherited.
+  size_t getFileSize() override;
+
+  // Documentation inherited.
+  size_t tell() override;
+
+  // Documentation inherited.
+  bool seek(ptrdiff_t _origin, SeekType _mode) override;
+
+  // Documentation inherited.
+  size_t read(void *_buffer, size_t _size, size_t _count) override;
+
+private:
+  std::FILE* mFile;
 };
 
-typedef std::shared_ptr<LocalResourceRetriever> LocalResourceRetrieverPtr;
-
-} // namespace utils
+} // namespace common
 } // namespace dart
 
-#endif // ifndef DART_UTILS_LOCALRESOURCERETRIEVER_H_
+#endif // ifndef DART_COMMON_LOCALRESOURCERETRIEVER_H_
