+++ dart/utils/urdf/DartLoader.cpp
@@ -38,7 +38,7 @@ namespace dart {
 namespace utils {
 
 DartLoader::DartLoader()
-  : mLocalRetriever(new utils::LocalResourceRetriever)
+  : mLocalRetriever(new common::LocalResourceRetriever)
   , mPackageRetriever(new utils::PackageResourceRetriever(mLocalRetriever))
   , mRetriever(new utils::SchemaResourceRetriever)
 {
